Crate x_core

Modules

core_config

git

Structs

WorkspaceSubset

Information collected about a subset of members of a workspace.

WorkspaceSubsets

Contains information about all the subsets specified in this workspace.

XCoreContext

Core context shared across all of x.

Enums

SystemError

A system error that happened while running a lint.

WorkspaceStatus

The status of a particular package ID in a WorkspaceSubset.

Type Definitions

Result

Type alias for the return type for run methods.